Sort squares in a consistent manner
This commit sorts squares in a consistent manner, favoring by values in
the y axis, and then on the x axis.

This does so by grouping by y values first:

    {0 => [square, square], 1 => [square, square]}

It then sorts each array subset along the x axis.

With each array sorted, we can then merge all the maps, extract their
values, and then flatten the resulting list.

I'd originally tried Enum.sort, but it was returning inconsistent
results because it was factoring both x and y values into the result.
This explicitly forces sorting across both axes independently and in the
correct order all the time. Is there a better way to accomplish this? I
hope so - I just haven't figured out how to do it yet.